divide packages by homework.
