+++ src/main/java/com/github/kirvasilchenko/t1/java/pro/homework1/AdvancedTestSuite.java
@@ -1,11 +1,11 @@
-package com.github.kirvasilchenko.t1.java.pro;
+package com.github.kirvasilchenko.t1.java.pro.homework1;
 
-import com.github.kirvasilchenko.t1.java.pro.annotation.AfterSuite;
-import com.github.kirvasilchenko.t1.java.pro.annotation.AfterTest;
-import com.github.kirvasilchenko.t1.java.pro.annotation.BeforeSuite;
-import com.github.kirvasilchenko.t1.java.pro.annotation.BeforeTest;
-import com.github.kirvasilchenko.t1.java.pro.annotation.CsvSource;
-import com.github.kirvasilchenko.t1.java.pro.annotation.Test;
+import com.github.kirvasilchenko.t1.java.pro.homework1.annotation.AfterSuite;
+import com.github.kirvasilchenko.t1.java.pro.homework1.annotation.AfterTest;
+import com.github.kirvasilchenko.t1.java.pro.homework1.annotation.BeforeSuite;
+import com.github.kirvasilchenko.t1.java.pro.homework1.annotation.BeforeTest;
+import com.github.kirvasilchenko.t1.java.pro.homework1.annotation.CsvSource;
+import com.github.kirvasilchenko.t1.java.pro.homework1.annotation.Test;
 
 public class AdvancedTestSuite {
