+++ src/main/java/com/github/kirvasilchenko/t1/java/pro/homework1/BasicTestSuite.java
@@ -1,8 +1,8 @@
-package com.github.kirvasilchenko.t1.java.pro;
+package com.github.kirvasilchenko.t1.java.pro.homework1;
 
-import com.github.kirvasilchenko.t1.java.pro.annotation.AfterSuite;
-import com.github.kirvasilchenko.t1.java.pro.annotation.BeforeSuite;
-import com.github.kirvasilchenko.t1.java.pro.annotation.Test;
+import com.github.kirvasilchenko.t1.java.pro.homework1.annotation.AfterSuite;
+import com.github.kirvasilchenko.t1.java.pro.homework1.annotation.BeforeSuite;
+import com.github.kirvasilchenko.t1.java.pro.homework1.annotation.Test;
 
 public class BasicTestSuite {
