+++ src/main/java/com/github/kirvasilchenko/t1/java/pro/homework1/annotation/AfterSuite.java
@@ -1,14 +1,13 @@
-package com.github.kirvasilchenko.t1.java.pro.annotation;
+package com.github.kirvasilchenko.t1.java.pro.homework1.annotation;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * Annotation for test. Set it in methods that are test cases.
+ * Annotation for test. Set it in static methods that should be executed after suite.
  */
 @Retention(RetentionPolicy.RUNTIME)
 @Target({java.lang.annotation.ElementType.METHOD})
-public @interface Test {
-    int priority() default 0;
+public @interface AfterSuite {
 }
