+++ src/main/java/com/github/kirvasilchenko/t1/java/pro/homework1/annotation/BeforeTest.java
@@ -1,13 +1,14 @@
-package com.github.kirvasilchenko.t1.java.pro.annotation;
+package com.github.kirvasilchenko.t1.java.pro.homework1.annotation;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * Annotation for test. Set it in static methods that should be executed after suite.
+ * Annotation for test. Set it in static methods that should be executed before each test.
  */
 @Retention(RetentionPolicy.RUNTIME)
 @Target({java.lang.annotation.ElementType.METHOD})
-public @interface AfterSuite {
+public @interface BeforeTest {
+
 }
