+++ src/main/java/com/github/kirvasilchenko/t1/java/pro/homework1/annotation/Test.java
@@ -1,14 +1,14 @@
-package com.github.kirvasilchenko.t1.java.pro.annotation;
+package com.github.kirvasilchenko.t1.java.pro.homework1.annotation;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * Annotation for test. Set it in static methods that should be executed before each test.
+ * Annotation for test. Set it in methods that are test cases.
  */
 @Retention(RetentionPolicy.RUNTIME)
 @Target({java.lang.annotation.ElementType.METHOD})
-public @interface BeforeTest {
-
+public @interface Test {
+    int priority() default 0;
 }
